Change definition of arithemtic operations on Sequences in 6.1
Arithmetic operations are now defined simply pointwise over
all indices, and starting index is the minimum over the starting
indices of the two operands. This pushes the burden of checking
whether an index is less or over the starting index to the user of the
sequence, without complicating the unfolding the arithmetic operations.

Also, add _eval theorems annotated with @simp.

Had to change the theorems in 6_5 to pass with the new definitions.